feat: add equations in markdown
The equations for the SNR calculation were added on top of the SNR function as a markdown.
The equation was simplified as much as possible, but it was not implemented that way yet.
I started from the equation avaliable in the excel sheet and simplified as much as I could.